feat(mres): absorb padding in single call
Conceptually a lot simpler, and it's not clear that huge amounts of padding would ever be used.

BREAKING CHANGE: Invalidates all previous Veil messages.

+++ src/main.rs
@@ -132,7 +132,7 @@ struct EncryptArgs {
 
     /// Add random bytes of padding.
     #[clap(long)]
-    padding: Option<u64>,
+    padding: Option<usize>,
 
     /// The path to read the passphrase from.
     #[clap(long, value_hint = ValueHint::FilePath)]
